// apps/mobile-technician/src/core/navigation/types.ts
export type RootStackParamList = {
  Login: undefined;
  Main: undefined;
};

export type MainTabParamList = {
  Home: undefined;
  Jobs: undefined;
  Notifications: undefined;
  Profile: undefined;
};

export type JobsStackParamList = {
  WorkOrders: undefined;
  WorkOrderDetail: undefined;
  TaskModules: undefined;
  BodyInspection: undefined;
  Evidence: undefined;
  FinalReview: undefined;
};